Add files via upload
Please type specifically for any and all input areas. Specifically, if you are asked a yes or no question, type in specifics yes or no in exact spelling. In addition names are specifically based on how they are written from the start. So, if you're asked to type a name, type that name from how you wrote it at the beginning. Not doing these things, such as typing "YES" instead of "yes" or "jimmy" instead of "Jimmy" if you had originally put it as "Jimmy" in the beginning will result in errors, meaning you will have to start the code from the beginning and it will be an absolute pain in the ass. So please, be specific.